Refactor training scripts
